Added a PROplot and moved plotting helper functions there for clarity
